Refactor MetadataHookForm into separate folder, refactor into hook
